Propagate the winner as a part of the saved URL
@elifkoc I don't believe this is actually wise--I would prefer to recover the winner from
the scores alone, which means that there's less for people to do, and the URL hash can be
shorter (because we're only storing two values per game, not three). But I would suggest
that the best way to figure this out is to schedule some time to talk it over with Frank
ASAP, and use him as a sounding board to figure out what will be usable.